Add tests for Thrust GPU and CPU backends
* Added decorator for skipping test based on unavailable method
* Add tests for Thrust GPU/CPU statevector simulator methods
* Add tests for Thrust GPU/CPU unitary simulator methods
* Add tests for Thrust GPU/CPU qasm simulator methods
* Renamed test files to be consistent
