+++ test/terra/backends/test_qasm_simulator_density_matrix.py
@@ -9,13 +9,13 @@
 # Any modifications or derivative works of this code must retain this
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
-
 """
 QasmSimulator Integration Tests
 """
 
 import unittest
 from test.terra import common
+from test.terra.decorators import requires_method
 
 # Basic circuit instruction tests
 from test.terra.backends.qasm_simulator.qasm_reset import QasmResetTests
@@ -53,40 +53,44 @@ from test.terra.backends.qasm_simulator.qasm_snapshot import QasmSnapshotExpValP
 from test.terra.backends.qasm_simulator.qasm_snapshot import QasmSnapshotExpValMatrixTests
 
 
-class TestQasmDensityMatrixSimulator(common.QiskitAerTestCase,
-                                     QasmMethodTests,
-                                     QasmMeasureTests,
-                                     QasmMultiQubitMeasureTests,
-                                     QasmResetTests,
-                                     QasmConditionalGateTests,
-                                     QasmConditionalUnitaryTests,
-                                     QasmConditionalKrausTests,
-                                     QasmConditionalSuperOpTests,
-                                     QasmCliffordTests,
-                                     QasmCliffordTestsWaltzBasis,
-                                     QasmCliffordTestsMinimalBasis,
-                                     QasmNonCliffordTests,
-                                     QasmNonCliffordTestsWaltzBasis,
-                                     QasmNonCliffordTestsMinimalBasis,
-                                     QasmAlgorithmTests,
-                                     QasmAlgorithmTestsWaltzBasis,
-                                     QasmAlgorithmTestsMinimalBasis,
-                                     QasmUnitaryGateTests,
-                                     QasmReadoutNoiseTests,
-                                     QasmPauliNoiseTests,
-                                     QasmResetNoiseTests,
-                                     QasmKrausNoiseTests,
-                                     QasmSnapshotStatevectorTests,
-                                     QasmSnapshotDensityMatrixTests,
-                                     QasmSnapshotProbabilitiesTests,
-                                     QasmSnapshotExpValPauliTests,
-                                     QasmSnapshotExpValMatrixTests,
-                                     QasmSnapshotStabilizerTests):
-    """QasmSimulator density_matrix method tests."""
+class DensityMatrixTests(
+        QasmMethodTests, QasmMeasureTests, QasmMultiQubitMeasureTests,
+        QasmResetTests, QasmConditionalGateTests, QasmConditionalUnitaryTests,
+        QasmConditionalKrausTests, QasmConditionalSuperOpTests,
+        QasmCliffordTests, QasmCliffordTestsWaltzBasis,
+        QasmCliffordTestsMinimalBasis, QasmNonCliffordTests,
+        QasmNonCliffordTestsWaltzBasis, QasmNonCliffordTestsMinimalBasis,
+        QasmAlgorithmTests, QasmAlgorithmTestsWaltzBasis,
+        QasmAlgorithmTestsMinimalBasis, QasmUnitaryGateTests,
+        QasmReadoutNoiseTests, QasmPauliNoiseTests, QasmResetNoiseTests,
+        QasmKrausNoiseTests, QasmSnapshotStatevectorTests,
+        QasmSnapshotDensityMatrixTests, QasmSnapshotProbabilitiesTests,
+        QasmSnapshotExpValPauliTests, QasmSnapshotExpValMatrixTests,
+        QasmSnapshotStabilizerTests):
+    """Container class of density_matrix method tests."""
+    pass
 
+
+class TestQasmSimulatorDensityMatrix(common.QiskitAerTestCase,
+                                     DensityMatrixTests):
+    """QasmSimulator density_matrix method tests."""
+    BACKEND_OPTS = {"seed_simulator": 314159, "method": "density_matrix"}
+
+
+@requires_method("qasm_simulator", "density_matrix_gpu")
+class TestQasmSimulatorDensityMatrixThrustGPU(common.QiskitAerTestCase,
+                                              DensityMatrixTests):
+    """QasmSimulator density_matrix_gpu method tests."""
+    BACKEND_OPTS = {"seed_simulator": 314159, "method": "density_matrix_gpu"}
+
+
+@requires_method("qasm_simulator", "density_matrix_thrust")
+class TestQasmSimulatorDensityMatrixThrustCPU(common.QiskitAerTestCase,
+                                              DensityMatrixTests):
+    """QasmSimulator density_matrix_thrust method tests."""
     BACKEND_OPTS = {
         "seed_simulator": 314159,
-        "method": "density_matrix"
+        "method": "density_matrix_thrust"
     }
