+++ test/terra/backends/test_qasm_simulator_extended_stabilizer.py
