+++ test/terra/backends/test_qasm_simulator_stabilizer.py
